Removed unused options according to lp:1300142
iscsi_helper this is moved to cinder in Grizzly. can be removed
connection_type this is deprecated in Folsom and removed in Grizzly. can be removed
root_helper this is deprecated in Folsom and removed in Grizzly. can be removed
volumes_path not sure when it deprecated, but removed before Grizzly.