set my_ip to unit private-address
[cbjchen,r=]
The default value for my_ip can be problematic.
The ip assocated with the NIC leading to the
default gateway may not be the desired one for
openstack service communication. Setting it to
the unit private-address, so that all the inter
service communications use the same network -
we alreay use the unit private-address as
service endpoint IPs.