Restart nova-api-metadata when nova.conf changes
The charm code as it stood before did this correctly when the
neutron-plugin relation included 'metadata-shared-secret', but not when
it included 'enable-metadata' without 'metadata-shared-secret'.

(See https://bugs.launchpad.net/charms/+source/nova-compute/+bug/1515570
for when 'enable-metadata' is used in this way.)

This change commonizes the code that determines when nova-api-metadata
should run, and uses this both to install the package and to add the
resource map dependency from nova.conf to the nova-api-metadata service.

Change-Id: I3354051b454621bc423ec6b2853d61301e58658c