Delete libvirt-bin default network to avoid MASQUERADE rules
libvirt-bin installs a 192.168.122.0/24 default network and creates
MASQUERADE rules for it on boot.  These rules will effect and break
instance traffic including GRE tenant networks.

Check if the network exists and then destroy it with virsh net-destroy
which both immediately removes the MASQUERADE rules and the network so
it is not applied after reboot.

Change-Id: Ia79aea6ef889d1ef58f903f967bea37dc07fd160
Closes-Bug: #1387390