Fix nova-compute charm with libvirt UCA
The UCA cloud archive recently added a newer version of
libvirt, however the daemon was not running after the charm
had finished installing. This was due to the fact that
the 'libvirtd' user was being selected based on the
Ubuntu release and not the openstack release.

Detect the openstack release as well that we are using
so we configure the libvirtd.conf properly.

Change-Id: Iffa2a00f12a252bd9eea76dde6f262779bc67158
Signed-off-by: Chuck Short <chuck.short@canonical.com>