Make the cloud-compute relation net spaces aware
When nova compute is deployed with multiple network interfaces
configured. It may be necessary to specify which interface to use for
the cloud-compute relation.

Do not use unit_get('private-address') which gives unpredictable
results. Instead, leverage network-get for network spaces aware address
selection.

Charm helpers sync to pull in generalized get_relation_ip() used in
other charms which checks network-get after checking for all of the edge
cases including IPv6 and config overrides. Use get_relation_ip() for
address selection.

This allows for specifying a network space in a bundle when using MAAS
2.x. This guarantees the charm will select the correct interface and IP
for the network space.

nova-compute:
  bindings:
    cloud-compute: internal-space

Closes-bug: #1670866

Change-Id: Ib43a7a52acf5a07b68dea808082da0ba6eb237c1